refactor: 파일 처리 유틸 리팩터링
- 파일 리스트 유틸과 파일 맵 유틸 하나로 병합
- 파일 관련 유틸 이름을 fileUtils로 변경
- 모듈 초기화 시 마크다운, 이미지 파일 맵 로드하고 요청 시 모듈에 로드된 맵 반환
